Support running as a shared library (#8)
* swap out HeapFactory with MemRegionAllocator

* rm typo

* add a real MemRegionAllocator

* fix accidental reinitialization of small/large heaps

* use std::optional instead of new() operator

* fix heap initialization typo

* log malloc ops

* move logging into allocator_interface.h

* add a macro for debug logging and a couple extra sanity checks

* fix log copy

* add cc file for allocator_interface

* rename globals

* fix BUILD type

* rm pragma once in cc file

* swap new() malloc flub

* swap new() malloc flub

* flush after _every_ log

* update logging macros to only flush if requested, they were tooo slow

* fix log condition

* fix log condition

* include thhread id in log

* hack to support new(0)

* disable debug logging when NDEBUG is defined

* remove redundant ERROR: prefix

* always return a pointer from new()

+++ src/jsmalloc/util/twiddle.h
@@ -2,12 +2,13 @@
 
 #include <cstddef>
 #include <cstdint>
+
 #include "src/jsmalloc/util/math.h"
 
 namespace jsmalloc {
 namespace twiddle {
 
-inline uint64_t PtrValue(const void* ptr) {
+inline intptr_t PtrValue(const void* ptr) {
   return reinterpret_cast<const uint8_t*>(ptr) - static_cast<uint8_t*>(nullptr);
 }
 
@@ -28,7 +29,7 @@ static inline constexpr T* OwnerOf(const M* ptr, const M T::*member) {
 }
 
 inline void* Align(void* ptr, size_t alignment) {
-  size_t val = PtrValue(ptr);
+  intptr_t val = PtrValue(ptr);
   return AddPtrOffset<void>(ptr, (math::round_up_pow2(val, alignment) - val));
 }
